Test 3201 tuning:
Omit the `throws FactoryException` statements because that exception is not thrown by the factory tested in this class.
This is because Units Of Measurement use a factory different than all other tested objects, provided by JSR-363 instead of GeoAPI.
Since `Test3201` is the only test case in that situation, it may not be worth to modify `TestGenerator` only for this test case.
Instead the `Test3201` file has been edited manually.

Since tests got manual editions, an additional simple test has been opportunistically added for some (not all) units.

The `baseUnit` and `unitToBase` fields have been made public for consistency with other tests.
This is controversial (a good practice would be to define getters and setters instead),
but we want to keep the test classes relatively simples.
Note: public access is for projects wanting to amend the tests.
As a compromise we could restrict to `protected` access, but we should do that for all classes.
that would be a separated issue.