Expand the 3205 tests for testing also geocentric and geographic 3D CRS in addition of geographic 2D. Add more verifications (e.g. axis lengths) about the datum by reusing the checks done by test 3204.
Note: the `csCode` field has been made public for consistency with current practice in other test files.
This is a practice that can be questioned if considered too bad practice; for now the intent is only to be consistent.
The current policy is to make public the fields containing information parsed from columns in the GIGS test file.
All other fields used for internal working of the test have more restricted access.